fix: Set "global.imageRegistry" to correct default value
The default is still our internal registry for now. Once the new public registry
has been set this will have to be updated again.

This does also set the value "image.registry" interim. There are pending tooling
adjustments before this will be changed into the empty string in the future.