test: Split ldap related test configuration into two separate modules
The chart is special in the sense that it does have multiple workload resources:

- Deployment
- Job
- CronJob

The idea is to run the client related tests per workload which uses the client.
This has the drawback of double checking some aspects of the way how the secret
is generated. The benefit is that the tests describe per workload the expected
behavior and can this way better serve as documentation of the intended
behavior.