fix: explicitly install python3-certifi
if not explicitly installed SBOM generation using trivy picks up the
certifi packages as a pypi dependency while it is actually the
debian package copied from a previous stage.

The debian package of python3-certifi is patched to use the debian
ca certificate bundle and fixes certain vulnerabilities present in
this version of the pypi certifi package.

Issue: univention/dev/internal/team-nubus#1540